Update README.md
Amend instruction to explicitly direct build path as of inside of the astro build config, not the dokploy settings. I could not get my project to deploy with the current wording.  Thanks to existing question from  Catur Hidayat in the discord in regards to a next.js deploy and a the solution from Siumauricio.